Check if input resolution is divisible by 2 before low resolution encoding

+++ backend/StreamServerApp/media_processing.py
@@ -183,7 +183,11 @@ def prepare_video(video_full_path,
         24 * 4 * (high_layer_compression_ratio/100.0)
     print("high_layer_bitrate = {}".format(high_layer_bitrate))
     low_layer_bitrate = int(os.getenv('480P_LAYER_BITRATE', 400000))
-    low_layer_height = int(video_height / 2.0)
+    low_layer_height = 0
+    if ((video_height % 2.0) == 0):
+        #We only encode low layer if lower resolution has the same aspect ratio as input
+        low_layer_height = int(video_height / 2.0)
+
     video_elementary_stream_path_low_layer = "{}_low.264".format(
         os.path.splitext(video_full_path)[0])
 
@@ -191,7 +195,7 @@ def prepare_video(video_full_path,
         video_full_path,
         video_elementary_stream_path_high_layer, video_height, high_layer_bitrate)
 
-    if low_layer_bitrate > 0:
+    if (low_layer_bitrate > 0 and low_layer_height > 0):
         h264_encoder(
             video_full_path,
             video_elementary_stream_path_low_layer, low_layer_height, low_layer_bitrate)
@@ -214,7 +218,7 @@ def prepare_video(video_full_path,
                   dash_output_directory)
 
     os.remove(video_elementary_stream_path_high_layer)
-    if low_layer_bitrate > 0:
+    if (low_layer_bitrate > 0 and low_layer_height > 0):
         os.remove(video_elementary_stream_path_low_layer)
 
     if not keep_files:
